Fix D.4 PMI scaling, polish plots, clean deps
The SN objective fits vectors v_SN = v̂/√(2d), so the correct prediction
is ⟨v_w, v_w'⟩ ≈ PMI (paper eq. 1.1), not ⟨v_w, v_w'⟩/d. This was the
headline diagnostic and was off by a factor of d (fitted slope 85 → 1.07
after the fix at d=300 / 100 epochs).

- D.4 scaling fix (src/verify.jl): drop /d from pmi_pred; expand docstring
  explaining the v_SN = v̂/√(2d) reparametrisation that absorbs the factor
  of d in Theorem 2.2
- Plot polish (scripts/04_verify.jl): D.1 title reports σ/𝔼 and fraction
  in [0.9, 1.1]; D.3 adds RMS/3 and observed-min reference lines; D.4 adds
  slope-1 reference y = x + γ (γ ≈ 3.81 for q=10) and _fit_line helper;
  D.6 replaces offset-by-0.35 hack with proper dodge-grouped bars + Legend
- SDE polish (scripts/06_sde_demo.jl): plot title includes σ_Z/Z̄; norm-
  preservation assertion (max |‖c‖ − 1| < 1e-10) guards the script
- Remove dead deps
- Notebook (notebooks/walkthrough.jl): correct the headline equation
  from PMI ≈ ⟨v,v'⟩/d to PMI ≈ ⟨v,v'⟩ (SN units, slope 1) in the
  model-slide cell, D.4 header, and D.4 plot xlabel; replace the
  continuous-time cell's Itô SDE + StochasticDiffEq.EM() description
  with the actual Stratonovich + projected Euler implementation
- Code quality (src/cooccurrence.jl, src/corpus.jl): PairDict value type
  Int32 → Int to avoid saturation on large corpora; remove hardcoded
  sizehint!(out, 17_000_000) from tokenize
- d=300 / 100-epoch retrain: wMSE 0.39 → 0.093; D.1 σ/𝔼 0.155 → 0.069
  (85% in [0.9, 1.1]); D.2 r 0.54 → 0.67; D.4 slope 0.85 → 1.07;
  SDE σ_Z/Z̄ 0.146 → 0.070; 59/59 tests pass

+++ src/cooccurrence.jl
@@ -1,7 +1,10 @@
 using OhMyThreads: tmapreduce, chunks
 using SparseArrays: SparseMatrixCSC, sparse
 
-const PairDict = Dict{Tuple{Int32, Int32}, Int32}
+# `Int` value type so the counter doesn't risk silent overflow on corpora
+# larger than text8 (Int32 saturates at ~2.1e9). `_expand_symmetric` still
+# materialises a `Float32` sparse matrix, so the public output is unchanged.
+const PairDict = Dict{Tuple{Int32, Int32}, Int}
 
 # Counts unordered pairs (i, j) with i < j whose token-stream positions lie in
 # `range × (range + window)`. Each unordered pair is therefore counted exactly
@@ -16,7 +19,7 @@ function _count_in_chunk(tokens::Vector{Int32}, range, window::Integer, N::Integ
             wj = tokens[j]
             wi == wj && continue
             key = wi < wj ? (wi, wj) : (wj, wi)
-            counts[key] = get(counts, key, Int32(0)) + Int32(1)
+            counts[key] = get(counts, key, 0) + 1
         end
     end
     return counts
